% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{bonferroni_alpha}
\alias{bonferroni_alpha}
\title{Bonferroni correction}
\usage{
bonferroni_alpha(alpha = 0.05, m)
}
\arguments{
\item{alpha}{Original significance level}

\item{m}{Number of tests}
}
\value{
Adjusted alpha level
}
\description{
Simple Bonferroni correction: adjust alpha by dividing by number of tests
}
