% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_helper.R
\name{cohens_d}
\alias{cohens_d}
\title{Calculate Cohen's d effect size}
\usage{
cohens_d(x, mu0 = 0)
}
\arguments{
\item{x}{Numeric vector of observations}

\item{mu0}{Null hypothesis value}
}
\value{
Numeric Cohen's d value
}
\description{
Computes standardized effect size for one-sample t-test
}
