% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{effect_size_t_test}
\alias{effect_size_t_test}
\title{Calculate detectable effect size for given power}
\usage{
effect_size_t_test(
  n,
  power = 0.8,
  sd,
  alpha = 0.05,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{n}{Sample size}

\item{power}{Desired power (default 0.80)}

\item{sd}{Standard deviation}

\item{alpha}{Significance level (default 0.05)}

\item{alternative}{Alternative hypothesis}
}
\value{
Detectable effect size (in original units)
}
\description{
Determines the effect size that can be detected with specified power
}
\examples{
# Effect size detectable with 80\% power and n=30
effect_size_t_test(n = 30, power = 0.80, sd = 1)
}
