% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visualizations.R
\name{ggplot_bootstrap}
\alias{ggplot_bootstrap}
\title{Bootstrap distribution plot}
\usage{
ggplot_bootstrap(bootstrap_result, theme = ggplot2::theme_minimal())
}
\arguments{
\item{bootstrap_result}{oneMeanTest_bootstrap object}

\item{theme}{ggplot2 theme}
}
\value{
ggplot object
}
\description{
Bootstrap distribution plot
}
