% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visualizations.R
\name{ggplot_ci}
\alias{ggplot_ci}
\title{Confidence interval plot with ggplot2}
\usage{
ggplot_ci(test_result, theme = ggplot2::theme_minimal())
}
\arguments{
\item{test_result}{oneMeanTest object}

\item{theme}{ggplot2 theme}
}
\value{
ggplot object
}
\description{
Confidence interval plot with ggplot2
}
