% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visualizations.R
\name{ggplot_distribution}
\alias{ggplot_distribution}
\title{Distribution plot with ggplot2}
\usage{
ggplot_distribution(x, test_result = NULL, theme = ggplot2::theme_minimal())
}
\arguments{
\item{x}{Numeric vector}

\item{test_result}{oneMeanTest object (optional)}

\item{theme}{ggplot2 theme}
}
\value{
ggplot object
}
\description{
Distribution plot with ggplot2
}
