% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visualizations.R
\name{ggplot_effect_sizes}
\alias{ggplot_effect_sizes}
\title{Effect size comparison plot}
\usage{
ggplot_effect_sizes(effect_size_summary, theme = ggplot2::theme_minimal())
}
\arguments{
\item{effect_size_summary}{Effect size summary object}

\item{theme}{ggplot2 theme}
}
\value{
ggplot object
}
\description{
Effect size comparison plot
}
