% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visualizations.R
\name{ggplot_multiple_tests}
\alias{ggplot_multiple_tests}
\title{Multiple tests comparison plot}
\usage{
ggplot_multiple_tests(multiple_tests_result, theme = ggplot2::theme_minimal())
}
\arguments{
\item{multiple_tests_result}{multiple_t_tests object}

\item{theme}{ggplot2 theme}
}
\value{
ggplot object
}
\description{
Multiple tests comparison plot
}
