% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visualizations.R
\name{ggplot_power_curve}
\alias{ggplot_power_curve}
\title{Power curve plot with ggplot2}
\usage{
ggplot_power_curve(
  delta,
  sd,
  alpha = 0.05,
  n_range = 10:100,
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{delta}{Effect size}

\item{sd}{Standard deviation}

\item{alpha}{Significance level}

\item{n_range}{Range of sample sizes}

\item{theme}{ggplot2 theme}
}
\value{
ggplot object
}
\description{
Power curve plot with ggplot2
}
