% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_visualizations.R
\name{ggplot_qq}
\alias{ggplot_qq}
\title{Q-Q plot with ggplot2}
\usage{
ggplot_qq(x, theme = ggplot2::theme_minimal())
}
\arguments{
\item{x}{Numeric vector}

\item{theme}{ggplot2 theme}
}
\value{
ggplot object
}
\description{
Q-Q plot with ggplot2
}
