% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_helper.R
\name{interpret_assumptions}
\alias{interpret_assumptions}
\title{Generate assumption check interpretation}
\usage{
interpret_assumptions(assumptions)
}
\arguments{
\item{assumptions}{Object from check_assumptions()}
}
\value{
Character string with interpretation
}
\description{
Interprets results of assumption checks
}
