% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_helper.R
\name{interpret_effect_size}
\alias{interpret_effect_size}
\title{Interpret effect size}
\usage{
interpret_effect_size(d)
}
\arguments{
\item{d}{Cohen's d effect size (standardized mean difference)}
}
\value{
Character description of effect size magnitude
}
\description{
Categorizes effect size using Cohen's conventions
}
