% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_helper.R
\name{interpret_p_value}
\alias{interpret_p_value}
\title{Interpret p-value strength}
\usage{
interpret_p_value(p_value)
}
\arguments{
\item{p_value}{Numeric p-value}
}
\value{
Character description of evidence strength
}
\description{
Categorizes p-value to describe strength of evidence
}
