% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_helper.R
\name{interpret_result}
\alias{interpret_result}
\title{Interpretation Helper Functions}
\usage{
interpret_result(result)
}
\arguments{
\item{result}{Object of class oneMeanTest}
}
\value{
Character string with interpretation
}
\description{
Functions to generate human-readable interpretations of statistical results
These provide context and plain-language explanations of test outcomes
}
\keyword{Creates}
\keyword{Generate}
\keyword{a}
\keyword{comprehensive}
\keyword{from}
\keyword{internal}
\keyword{interpretation}
\keyword{of}
\keyword{plain-language}
\keyword{result}
\keyword{results}
\keyword{test}
