% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric.R
\name{nonparametric}
\alias{nonparametric}
\title{Non-Parametric Alternatives for One-Sample Location Test}
\description{
Functions to perform Wilcoxon signed-rank test and other
non-parametric alternatives to the one-sample t-test
}
