% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneMeanTest-packager.R
\docType{package}
\name{oneMeanTest}
\alias{oneMeanTest-package}
\alias{oneMeanTest}
\title{oneMeanTest: One-sample t-test package}
\description{
The \pkg{oneMeanTest} package provides functions to perform a one-sample
t-test for a population mean when the population variance is unknown, along
with assumption checks, descriptive statistics, and basic plotting utilities
for educational use.
}
\author{
\strong{Maintainer}: Group 1 3L STAT181 \email{group1_3L@example.com}

}
