% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.oneMeanTest_bootstrap}
\alias{plot.oneMeanTest_bootstrap}
\title{Plot bootstrap distribution}
\usage{
\method{plot}{oneMeanTest_bootstrap}(x, ...)
}
\arguments{
\item{x}{oneMeanTest_bootstrap object}

\item{...}{Additional graphical parameters}
}
\description{
Plot bootstrap distribution
}
