% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Boxplot with outlier identification}
\usage{
plot_boxplot(x, ...)
}
\arguments{
\item{x}{Numeric vector}

\item{...}{Additional graphical parameters}
}
\description{
Boxplot with outlier identification
}
