% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ci}
\alias{plot_ci}
\title{Plot confidence interval}
\usage{
plot_ci(x, ...)
}
\arguments{
\item{x}{oneMeanTest object}

\item{...}{Additional graphical parameters}
}
\description{
Plot confidence interval
}
