% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot data distribution with test information}
\usage{
plot_distribution(x, test_obj = NULL, ...)
}
\arguments{
\item{x}{oneMeanTest object or numeric vector}

\item{test_obj}{oneMeanTest object (optional, when x is numeric data)}

\item{...}{Additional graphical parameters}
}
\description{
Plot data distribution with test information
}
