% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Histogram with normal curve}
\usage{
plot_histogram(x, ...)
}
\arguments{
\item{x}{Numeric vector}

\item{...}{Additional graphical parameters}
}
\description{
Histogram with normal curve
}
