% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_qq}
\alias{plot_qq}
\title{Q-Q plot for normality assessment}
\usage{
plot_qq(x, ...)
}
\arguments{
\item{x}{Numeric vector}

\item{...}{Additional graphical parameters}
}
\description{
Q-Q plot for normality assessment
}
