% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{print.effect_size}
\alias{print.effect_size}
\title{Print method for effect size objects}
\usage{
\method{print}{effect_size}(x, digits = 4, ...)
}
\arguments{
\item{x}{Effect size object}

\item{digits}{Number of decimal places (default 4)}

\item{...}{Additional arguments}
}
\description{
Print method for effect size objects
}
