% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sizes.R
\name{print.effect_size_summary}
\alias{print.effect_size_summary}
\title{Print method for effect size summary}
\usage{
\method{print}{effect_size_summary}(x, digits = 4, ...)
}
\arguments{
\item{x}{Effect size summary object}

\item{digits}{Number of decimal places (default 4)}

\item{...}{Additional arguments}
}
\description{
Print method for effect size summary
}
