% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{print.multiple_t_tests}
\alias{print.multiple_t_tests}
\title{Print method for multiple t-tests}
\usage{
\method{print}{multiple_t_tests}(x, digits = 4, ...)
}
\arguments{
\item{x}{Multiple t-tests object}

\item{digits}{Number of decimal places}

\item{...}{Additional arguments}
}
\description{
Print method for multiple t-tests
}
