% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{print.multiple_testing}
\alias{print.multiple_testing}
\title{Print method for multiple testing}
\usage{
\method{print}{multiple_testing}(x, digits = 4, ...)
}
\arguments{
\item{x}{Multiple testing object}

\item{digits}{Number of decimal places}

\item{...}{Additional arguments}
}
\description{
Print method for multiple testing
}
