% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tidy_methods.R
\name{print.oneMeanTest}
\alias{print.oneMeanTest}
\title{Print Method for oneMeanTest Objects}
\usage{
\method{print}{oneMeanTest}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class oneMeanTest}

\item{digits}{Number of digits to display}

\item{...}{Additional arguments (not currently used)}
}
\description{
Prints a formatted summary of one-sample t-test results
}
