% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{print.oneMeanTest_bootstrap}
\alias{print.oneMeanTest_bootstrap}
\title{Print method for bootstrap one-sample t-test}
\usage{
\method{print}{oneMeanTest_bootstrap}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{"oneMeanTest_bootstrap"}.}

\item{digits}{Number of digits to use for printing numeric values.}

\item{...}{Unused; included for method compatibility.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Pretty printing for objects of class \code{"oneMeanTest_bootstrap"} produced
by \code{\link{bootstrap_ttest}}.
}
