% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric.R
\name{print.oneMeanTest_wilcoxon}
\alias{print.oneMeanTest_wilcoxon}
\title{Print method for Wilcoxon test}
\usage{
\method{print}{oneMeanTest_wilcoxon}(x, digits = 4, ...)
}
\arguments{
\item{x}{Wilcoxon test object}

\item{digits}{Number of decimal places}

\item{...}{Additional arguments}
}
\description{
Print method for Wilcoxon test
}
