% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{sample_size_t_test}
\alias{sample_size_t_test}
\title{Calculate required sample size for desired power}
\usage{
sample_size_t_test(
  power = 0.8,
  delta,
  sd,
  alpha = 0.05,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{power}{Desired power (default 0.80)}

\item{delta}{True difference from null value}

\item{sd}{Standard deviation}

\item{alpha}{Significance level (default 0.05)}

\item{alternative}{Alternative hypothesis}
}
\value{
Required sample size (integer)
}
\description{
Determines the sample size needed to achieve specified power.
}
\examples{
# Sample size needed for 80\% power to detect effect of 0.5 SD
sample_size_t_test(power = 0.80, delta = 0.5, sd = 1)
}
