% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric.R
\name{sign_test}
\alias{sign_test}
\title{Sign Test}
\usage{
sign_test(x, mu0 = 0, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{x}{Numeric vector of observations}

\item{mu0}{Hypothesized median (default 0)}

\item{alternative}{Alternative hypothesis}
}
\value{
List with test results
}
\description{
Simple non-parametric test based on signs of differences
}
