% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tidy_methods.R
\name{summary.oneMeanTest}
\alias{summary.oneMeanTest}
\title{Summary Method for oneMeanTest Objects}
\usage{
\method{summary}{oneMeanTest}(object, ...)
}
\arguments{
\item{object}{An object of class oneMeanTest}

\item{...}{Additional arguments passed to print method}
}
\description{
Provides a summary of one-sample t-test results
}
