% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.oneMeanTest}
\alias{plot.oneMeanTest}
\title{Plot methods for one-sample mean test objects}
\usage{
\method{plot}{oneMeanTest}(x, which = c("t", "hist", "qq", "box", "ci"), ...)
}
\arguments{
\item{x}{An object of class \code{"oneMeanTest"}.}

\item{which}{Character string specifying which plot to draw. One of
\code{"t"}, \code{"hist"}, \code{"qq"}, \code{"box"}, or \code{"ci"}.}

\item{...}{Additional arguments passed on to the underlying graphics
functions (currently unused).}
}
\value{
Invisibly returns \code{x} after producing a plot.
}
\description{
Plot methods for objects of class \code{"oneMeanTest"} produced by
\code{\link{one_mean_test}}. Different values of \code{which} produce
different visualizations:
}
\details{
\itemize{
\item \code{"t"}: t-distribution with the observed test statistic and
critical values for the chosen \code{alpha}.
\item \code{"hist"}: histogram of the data with vertical lines for the
sample mean and hypothesized mean \code{mu0}.
\item \code{"qq"}: normal Q-Q plot of the data with a reference line.
\item \code{"box"}: boxplot of the data with a vertical line at \code{mu0}.
\item \code{"ci"}: confidence interval plot for the mean.
}

For \code{"hist"}, \code{"qq"}, and \code{"box"}, the raw data must be
stored in the object via \code{attr(x, "data") <- your_data}.
}
\examples{
set.seed(123)
x <- rnorm(30, mean = 5, sd = 2)
res <- one_mean_test(x, mu0 = 5, check_assumptions = FALSE)
attr(res, "data") <- x

plot(res, which = "t")
plot(res, which = "ci")
plot(res, which = "hist")
plot(res, which = "qq")
plot(res, which = "box")

}
