% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_methods.R
\name{report}
\alias{report}
\title{Report Method for oneMeanTest}
\usage{
report(x, format = "console", ...)
}
\arguments{
\item{x}{Object of class oneMeanTest}

\item{format}{Output format: "console", "markdown", or "latex"}
}
\description{
Report Method for oneMeanTest
}
